requirements.txt
Flask==2.2.5
Werkzeug==2.2.3
itsdangerous==2.1.2
Jinja2==3.0.3
MarkupSafe==2.1.3
Flask-SQLAlchemy==2.5.1
Flask-Login==0.6.2
Flask-Bcrypt==1.0.1
Flask-WTF==1.0.1
WTForms==3.0.1
python-dotenv==0.21.0
Flask-Mail==0.9.1
#psycopg2-binary==2.9.6  # Only needed for PostgreSQL
email-validator==1.3.1
gunicorn==20.1.0
pytest==7.2.2
pytest-flask==1.2.0
Flask-Migrate==3.1.0
alembic==1.7.7
SQLAlchemy==1.4.46
